bundle types correctly
