+++ npm/react/src/mount.ts
@@ -16,7 +16,6 @@ setupHooks()
  */
 const injectStyles = (options: MountOptions) => {
   return () => {
-    const document = cy.state('document')
     const el = document.getElementById(rootId)
 
     return injectStylesBeforeElement(options, document, el)
@@ -50,6 +49,8 @@ export const mount = (jsx: React.ReactNode, options: MountOptions = {}) => {
   const message = options.alias
     ? `<${componentName} ... /> as "${options.alias}"`
     : `<${componentName} ... />`
+
+  // @ts-ignore
   let logInstance: Cypress.Log
 
   return cy
@@ -63,7 +64,6 @@ export const mount = (jsx: React.ReactNode, options: MountOptions = {}) => {
   })
   .then(injectStyles(options))
   .then(() => {
-    const document = cy.state('document') as Document
     const reactDomToUse = options.ReactDom || ReactDOM
 
     const el = document.getElementById(rootId)
@@ -199,3 +199,123 @@ export const createMount = (defaultOptions: MountOptions) => {
 
 /** @deprecated Should be removed in the next major version */
 export default mount
+
+// I hope to get types and docs from functions imported from ./index one day
+// but for now have to document methods in both places
+// like this: import {mount} from './index'
+
+export interface ReactModule {
+  name: string
+  type: string
+  location: string
+  source: string
+}
+
+/**
+ * Additional styles to inject into the document.
+ * A component might need 3rd party libraries from CDN,
+ * local CSS files and custom styles.
+ */
+export interface StyleOptions {
+  /**
+   * Creates <link href="..." /> element for each stylesheet
+   * @alias stylesheet
+   */
+  stylesheets: string | string[]
+  /**
+   * Creates <link href="..." /> element for each stylesheet
+   * @alias stylesheets
+   */
+  stylesheet: string | string[]
+  /**
+   * Creates <style>...</style> element and inserts given CSS.
+   * @alias styles
+   */
+  style: string | string[]
+  /**
+   * Creates <style>...</style> element for each given CSS text.
+   * @alias style
+   */
+  styles: string | string[]
+  /**
+   * Loads each file and creates a <style>...</style> element
+   * with the loaded CSS
+   * @alias cssFile
+   */
+  cssFiles: string | string[]
+  /**
+   * Single CSS file to load into a <style></style> element
+   * @alias cssFile
+   */
+  cssFile: string | string[]
+}
+
+export interface MountReactComponentOptions {
+  alias: string
+  ReactDom: typeof ReactDOM
+  /**
+   * Log the mounting command into Cypress Command Log,
+   * true by default.
+   */
+  log: boolean
+  /**
+   * Render component in React [strict mode](https://reactjs.org/docs/strict-mode.html)
+   * It activates additional checks and warnings for child components.
+   */
+  strict: boolean
+}
+
+export type MountOptions = Partial<StyleOptions & MountReactComponentOptions>
+
+/**
+ * The `type` property from the transpiled JSX object.
+ * @example
+ * const { type } = React.createElement('div', null, 'Hello')
+ * const { type } = <div>Hello</div>
+ */
+export interface JSX extends Function {
+  displayName: string
+}
+
+export declare namespace Cypress {
+  interface Cypress {
+    stylesCache: any
+    React: string
+    ReactDOM: string
+    Styles: string
+    modules: ReactModule[]
+  }
+
+  // NOTE: By default, avoiding React.Component/Element typings
+  // for many cases, we don't want to import @types/react
+  interface Chainable<Subject> {
+    // adding quick "cy.state" method to avoid TS errors
+    state: (key: string) => any
+    // injectReactDOM: () => Chainable<void>
+    // copyCompon { ReactDom }entStyles: (component: Symbol) => Chainable<void>
+    /**
+     * Mount a React component in a blank document; register it as an alias
+     * To access: use an alias or original component reference
+     *  @function   cy.mount
+     *  @param      {Object}  jsx - component to mount
+     *  @param      {string}  [Component] - alias to use later
+     *  @example
+    ```
+    import Hello from './hello.jsx'
+    // mount and access by alias
+    cy.mount(<Hello />, 'Hello')
+    // using default alias
+    cy.get('@Component')
+    // using specified alias
+    cy.get('@Hello').its('state').should(...)
+    // using original component
+    cy.get(Hello)
+    ```
+    **/
+    // mount: (component: Symbol, alias?: string) => Chainable<void>
+    get<S = any>(
+      alias: string | symbol | Function,
+      options?: Partial<any>,
+    ): Chainable<any>
+  }
+}
